Library mode (external main) now can be enabled with build flag
