+++ cmd/ck-client/ck-client.go
@@ -10,8 +10,6 @@ import (
 	"fmt"
 	"net"
 	"os"
-	"C"
-	"unsafe"
 
 	"github.com/cbeuw/Cloak/internal/common"
 
@@ -19,19 +17,6 @@ import (
 	mux "github.com/cbeuw/Cloak/internal/multiplex"
 	log "github.com/sirupsen/logrus"
 )
-
-//export external_main
-func external_main(argc C.int, argv **C.char) {
-	var offset = unsafe.Sizeof(uintptr(0))
-	var go_argv []string
-	for i := 0; i < int(argc); i++ {
-		go_argv = append(go_argv, C.GoString(*argv))
-		argv = (**C.char)(unsafe.Pointer(uintptr(unsafe.Pointer(argv)) + offset))
-	}
-
-	os.Args = go_argv
-	main()
-}
 
 var version string
